requirements.txt
aiohttp>=3.8.0
asyncio>=3.4.3
bs4>=0.0.1
cachetools>=5.0.0
jinja2>=3.0.0
langchain>=0.0.267
langchain-google-genai>=0.0.6
langchain-openai>=0.0.2
loguru>=0.7.0
mcp>=0.0.6
python-dotenv>=1.0.0
requests>=2.28.0
langgraph>=0.0.20
langchain-mcp-adapters>=0.0.2
pydantic-settings>=2.0.0
brotli>=1.0.9
